refactor(ts): move runtime helpers to separate file
